requirements.txt
asgiref==3.4.1
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.9
django-compat==1.0.15
django-cors-headers==3.10.1
django-import-export==2.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
firebase-admin==5.2.0
gcloud==0.18.3
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
jwcrypto==1.0
MarkupPy==1.14
msgpack==1.0.3
ndg-httpsclient==0.5.1
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==8.4.0
proto-plus==1.19.8
protobuf==3.19.3
psycopg2==2.9.2
pusher==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-jwt==3.3.2
pytz==2021.3
PyYAML==6.0
reportlab==3.6.5
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
